Did a major regrouping of code
