+++ CMakeLists.txt
@@ -78,7 +78,9 @@ add_library(Delphi
   lib/Indicator.cpp
   lib/parameterization.cpp
   lib/subgraphs.cpp
-  lib/json_constructors.cpp
+  lib/constructors.cpp
+  lib/sampling.cpp
+  lib/inference.cpp
 )
 
 target_include_directories(Delphi PRIVATE external/range-v3/include)
