+++ tests/wm/test_cpp_evaluation.py
@@ -27,7 +27,6 @@ def test_cpp_extensions():
     # InitialBeta.HALF
     # InitialBeta.MEAN
     # InitialBeta.RANDOM - A random value between [-1, 1]
-    G.construct_beta_pdfs()
     G.train_model(
         2015, 1, 2015, 12, 1000, 10000, initial_beta=InitialBeta.ZERO
     )
